Moving Card component into utilities directory.
