+++ resources/assets/components/LegacyQuiz/Conclusion.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Card from '../Card';
+
 import Markdown from '../Markdown';
+import Card from '../utilities/Card/Card';
 
 const Conclusion = props => {
   const { children, callToAction } = props;
