+++ resources/assets/components/Page/LandingPage/templates/PitchTemplate.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Markdown from '../../../Markdown';
-import Card from '../../../Card';
+import Card from '../../../utilities/Card/Card';
 
 const PitchTemplate = ({ pitchContent, sidebarCTA }) => (
   <div className="campaign-subpage">
