+++ resources/assets/components/Quiz/QuizChoice.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-import Card from '../Card';
+import Card from '../utilities/Card/Card';
 
 const QuizChoice = props => {
   const {
