+++ resources/assets/components/ShareAction/ShareAction.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Card from '../Card';
 import Embed from '../Embed';
 import Markdown from '../Markdown';
+import Card from '../utilities/Card/Card';
 import { POST_SHARE_MODAL } from '../Modal';
 import { showFacebookSharePrompt, showTwitterSharePrompt } from '../../helpers';
