+++ resources/assets/components/utilities/Card/Card.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { shallowToJson } from 'enzyme-to-json';
+
 import Card from './Card';
 
 const component = shallow(
