+++ resources/assets/components/utilities/Card/card.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .card {
   background-color: $white;
